fix(nextly): abort batch on version conflict and gate reads on outbox
- Revert the batch-wide version-conflict retry: re-running the transaction
  replays every preceding item's non-transactional hooks (duplicate afterCreate
  effects, discarded ids). A conflict now aborts the batch cleanly — no orphan
  row or version, no hook replay — and surfaces as a failure the caller can
  retry. Creates never collide (unique ids, per-document sequences), so only
  concurrent same-entry updates are affected; a hook-safe retry (savepoint or
  serialized allocation) is left as a follow-up.
- Gate the tx-API/batch relational assembly on the full outbox gate
  (isOutboxRecordingActive) instead of the per-entity opt-out, so a
  webhook-permitted collection with no endpoint and audit off skips the
  component/m2m reads for a payload nobody records.

Replace the retry test with one asserting the clean abort leaves nothing behind.
